Make sure the file exists before reading it.
This is used for both haproxy and nginx. This commit has been tested on a
faststart image.

+++ AppController/djinn.rb
@@ -575,9 +575,7 @@ class Djinn
   # A SOAP-exposed method that callers can use to get information about what
   # roles each node in the AppScale deployment are running.
   def get_role_info(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     all_nodes = []
     @nodes.each { |node|
@@ -599,9 +597,7 @@ class Djinn
   #   can be authenticated, a JSON-dumped Hash containing information about
   #   applications on this machine is returned.
   def get_app_info_map(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     return JSON.dump(@app_info_map)
   end
@@ -623,9 +619,7 @@ class Djinn
   #   "OK" if the relocation occurred successfully, and a String containing the
   #   reason why the relocation failed in all other cases.
   def relocate_app(appid, http_port, https_port, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if !my_node.is_shadow?
       # We need to send the call to the shadow.
@@ -634,7 +628,7 @@ class Djinn
       begin
         return acc.relocate_app(appid, http_port, https_port)
       rescue FailedNodeException => except
-        Djinn.log_warn("Failed to forward relocate_app call to shadow (#{get_shadow})."
+        Djinn.log_warn("Failed to forward relocate_app call to shadow (#{get_shadow}).")
         return NOT_READY
       end
     end
@@ -749,8 +743,11 @@ class Djinn
   #   A String indicating that the termination has started, or the reason why it
   #   failed.
   def kill(stop_deployment, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
+    begin
+      return BAD_SECRET_MSG if !valid_secret?(secret)
+    rescue Errno::ENOENT
+      # On appscale down, terminate may delete our secret key before we
+      # can check it here.
     end
     @kill_sig_received = true
 
@@ -1059,7 +1056,7 @@ class Djinn
       begin
         return acc.upload_app(archived_file, file_suffix, email)
       rescue FailedNodeException => except
-        Djinn.log_warn("Failed to forward upload_app call to shadow (#{get_shadow})."
+        Djinn.log_warn("Failed to forward upload_app call to shadow (#{get_shadow}).")
         return NOT_READY
       end
     end
@@ -1299,9 +1296,7 @@ class Djinn
   #     - KEY_NOT_FOUND if there is no instance variable with the given name.
   #     - BAD_SECRET_MSG if the caller could not be authenticated.
   def set_property(property_name, property_value, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     Djinn.log_info("Attempting to set @#{property_name} to #{property_value}")
 
@@ -1321,9 +1316,7 @@ class Djinn
   # Returns:
   #   A boolean indicating whether the deployment ID has been set or not.
   def deployment_id_exists(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     return ZKInterface.exists?(DEPLOYMENT_ID_PATH)
   end
@@ -1332,9 +1325,7 @@ class Djinn
   # Returns:
   #   A string that contains the deployment ID.
   def get_deployment_id(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       return ZKInterface.get(DEPLOYMENT_ID_PATH)
@@ -1349,9 +1340,7 @@ class Djinn
   # Args:
   #   id: A string that contains the deployment ID.
   def set_deployment_id(secret, id)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       ZKInterface.set(DEPLOYMENT_ID_PATH, id, false)
@@ -1367,9 +1356,7 @@ class Djinn
   #   read_only: A string that indicates whether to turn read-only mode on or
   #     off.
   def set_node_read_only(read_only, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
     return INVALID_REQUEST unless %w(true false).include?(read_only)
     read_only = read_only == 'true'
 
@@ -1388,9 +1375,7 @@ class Djinn
   #   read_only: A string that indicates whether to turn read-only mode on or
   #     off.
   def set_read_only(read_only, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
     return INVALID_REQUEST unless %w(true false).include?(read_only)
 
     @nodes.each { | node |
@@ -1412,9 +1397,7 @@ class Djinn
   # Returns:
   #   A boolean indicating whether or not the user application exists.
   def does_app_exist(appname, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1431,9 +1414,7 @@ class Djinn
   #   username: The email address for the user whose password will be changed.
   #   password: The SHA1-hashed password that will be set as the user's password.
   def reset_password(username, password, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1449,9 +1430,7 @@ class Djinn
   # Args:
   #   username: The email address registered as username for the user's application.
   def does_user_exist(username, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1470,9 +1449,7 @@ class Djinn
   #   account_type: A str that indicates if this account can be logged into
   #     by XMPP users.
   def create_user(username, password, account_type, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1488,9 +1465,7 @@ class Djinn
   # Args:
   #   username: The e-mail address that should be given administrative authorizations.
   def set_admin_role(username, is_cloud_admin, capabilities, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1508,9 +1483,7 @@ class Djinn
   #  Returns:
   #    A JSON-encoded string containing the application metadata.
   def get_app_data(app_id, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1530,9 +1503,7 @@ class Djinn
   #   app_language: The runtime (Python 2.5/2.7, Java, or Go) that the app
   #     runs over.
   def reserve_app_id(username, app_id, app_language, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     begin
       uac = UserAppClient.new(my_node.private_ip, @@secret)
@@ -1550,9 +1521,7 @@ class Djinn
   #   secret: Shared key for authentication
   #
   def stop_app(app_name, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if !my_node.is_shadow?
       Djinn.log_debug("Sending stop_app call for #{app_name} to shadow.")
@@ -1560,7 +1529,7 @@ class Djinn
       begin
         return acc.stop_app(app_name)
       rescue FailedNodeException => except
-        Djinn.log_warn("Failed to forward stop_app call to shadow (#{get_shadow})."
+        Djinn.log_warn("Failed to forward stop_app call to shadow (#{get_shadow}).")
         return NOT_READY
       end
     end
@@ -1658,9 +1627,7 @@ class Djinn
   end
 
   def update(app_names, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     # Few sanity checks before acting.
     if !my_node.is_shadow?
@@ -1669,7 +1636,7 @@ class Djinn
        begin
          return acc.update(app_names)
        rescue FailedNodeException => except
-         Djinn.log_warn("Failed to forward update call to shadow (#{get_shadow})."
+         Djinn.log_warn("Failed to forward update call to shadow (#{get_shadow}).")
          return NOT_READY
        end
     end
@@ -1749,9 +1716,7 @@ class Djinn
   #   A String indicating that the SOAP call succeeded, or the reason why it
   #   did not.
   def set_apps_to_restart(apps_to_restart, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     APPS_LOCK.synchronize {
       @apps_to_restart += apps_to_restart
@@ -1763,9 +1728,7 @@ class Djinn
   end
 
   def get_all_public_ips(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     public_ips = []
     @nodes.each { |node|
@@ -1775,9 +1738,7 @@ class Djinn
   end
 
   def job_start(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     Djinn.log_info("==== Starting AppController (pid: #{Process.pid}) ====")
 
@@ -1949,9 +1910,7 @@ class Djinn
 
 
   def get_online_users_list(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     online_users = []
 
@@ -1967,9 +1926,7 @@ class Djinn
   end
 
   def done_uploading(appname, location, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if File.exists?(location)
       begin
@@ -1991,9 +1948,7 @@ class Djinn
   end
 
   def is_app_running(appname, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     hosters = ZKInterface.get_app_hosters(appname, @options['keyname'])
     hosters_w_appengine = []
@@ -2008,9 +1963,7 @@ class Djinn
 
 
   def add_role(new_role, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     # new roles may run indefinitely in the background, so don't block
     # on them - just fire and forget
@@ -2034,9 +1987,7 @@ class Djinn
   end
 
   def remove_role(old_role, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     my_node.remove_roles(old_role)
     stop_roles = old_role.split(":")
@@ -2066,9 +2017,7 @@ class Djinn
   #   Otherwise, returns a Hash that maps IP addresses to the roles that
   #     will be hosted on them (the inverse of ips_hash).
   def start_roles_on_nodes(ips_hash, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     ips_hash = JSON.load(ips_hash)
     if ips_hash.class != Hash
@@ -2195,9 +2144,7 @@ class Djinn
   # combination of the two. 'nodes_needed' should be an Array, where each
   # item is an Array of the roles to start on each node.
   def start_new_roles_on_nodes(nodes_needed, instance_type, secret)
-     if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if nodes_needed.class != Array
       Djinn.log_error("Was expecting nodes_needed to be an Array, not " +
@@ -2590,9 +2537,7 @@ class Djinn
   # Args:
   #   secret: A String password that is used to authenticate SOAP callers.
   def gather_logs(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     uuid = HelperFunctions.get_random_alphanumeric()
     Djinn.log_info("Generated uuid #{uuid} for request to gather logs.")
@@ -2645,9 +2590,7 @@ class Djinn
   #   - NOT_READY: If this node runs HAProxy, but hasn't allocated ports for
   #     it and nginx yet. Callers should retry at a later time.
   def add_routing_for_appserver(app_id, ip, port, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if !my_node.is_shadow?
        # We need to send the call to the shadow.
@@ -2656,7 +2599,7 @@ class Djinn
        begin
          return acc.add_routing_for_appserver(app_id, ip, port)
        rescue FailedNodeException => except
-         Djinn.log_warn("Failed to forward routing call to shadow (#{get_shadow})."
+         Djinn.log_warn("Failed to forward routing call to shadow (#{get_shadow}).")
          return NOT_READY
        end
     end
@@ -2695,9 +2638,7 @@ class Djinn
   #   - BAD_SECRET_MSG: If the caller cannot be authenticated.
   #   - NO_HAPROXY_PRESENT: If this node does not run HAProxy.
   def add_routing_for_blob_server(secret)
-    unless valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     unless my_node.is_login?
       return NO_HAPROXY_PRESENT
@@ -2734,9 +2675,7 @@ class Djinn
   #   - NO_HAPROXY_PRESENT: If this node does not run HAProxy (and thus cannot
   #     remove AppServers from HAProxy config files).
   def remove_appserver_from_haproxy(app_id, ip, port, secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     if !my_node.is_shadow?
        Djinn.log_debug("Sending remote_appserver_from_haproxy call for #{app_id} to shadow.")
@@ -4217,8 +4156,8 @@ HOSTS
     my_private = my_node.private_ip
     login_ip = get_login.private_ip
 
-    Djinn.log_debug("@app_info_map is #{@app_info_map.inspect}")
-    @apps_loaded.each { |app|
+    apps_to_check = @apps_loaded + RESERVED_APPS
+    apps_to_check.each { |app|
       next if app == "none"
 
       http_port = @app_info_map[app]['nginx']
@@ -4257,7 +4196,6 @@ HOSTS
       end
     }
     Djinn.log_debug("Done updating nginx and haproxy config files.")
-    Nginx.reload()
   end
 
 
@@ -4656,6 +4594,7 @@ HOSTS
   def check_stopped_app()
     HelperFunctions.get_apps_loaded().each{ |app|
       next if @app_names.include?(app)
+      next if RESERVED_APPS.include?(app)
 
       Djinn.log_info("#{app} is no longer running: removing old states.")
       Djinn.log_run("rm -rf #{HelperFunctions.get_app_path(app)}")
@@ -5930,9 +5869,7 @@ HOSTS
   #   A hash in string format containing system and platform stats for this
   #     node.
   def get_all_stats(secret)
-    if !valid_secret?(secret)
-      return BAD_SECRET_MSG
-    end
+    return BAD_SECRET_MSG if !valid_secret?(secret)
 
     # Get stats from SystemManager.
     imc = InfrastructureManagerClient.new(secret)
